Change setUpLTI to LTIService; handle Promise in constructor of LTIMidddleware
