+++ ccm_web/server/src/lti/lti.middleware.ts
@@ -1,32 +1,29 @@
-import { Injectable, NestMiddleware } from '@nestjs/common'
-import { ConfigService } from '@nestjs/config'
 import { Express, Request, Response, NextFunction } from 'express'
+import { Injectable, NestMiddleware } from '@nestjs/common'
 
-import { DatabaseConfig, LTIConfig } from '../config'
-import setUpLTI from './lti.setup'
+import { LTIService } from './lti.service'
+
+import baseLogger from '../logger'
+
+const logger = baseLogger.child({ filePath: __filename })
 
 @Injectable()
 export class LTIMiddleware implements NestMiddleware {
   ltiMiddleware: Express | undefined
 
-  constructor (private readonly configService: ConfigService) {}
-
-  public async setLTIMiddleware (): Promise<Express> {
-    const dbConfig = this.configService.get('db') as DatabaseConfig
-    const ltiConfig = this.configService.get('lti') as LTIConfig
-
-    const ltiMiddleware = await setUpLTI(dbConfig, ltiConfig)
-    this.ltiMiddleware = ltiMiddleware
-    return ltiMiddleware
+  constructor (private readonly ltiService: LTIService) {
+    this.ltiService.setUpLTI()
+      .then(app => {
+        this.ltiMiddleware = app
+        logger.info('ltijs was successfully set up.')
+      })
+      .catch(e => {
+        this.ltiMiddleware = undefined
+        throw new Error(`An error occurred while setting up ltijs: ${String(e)}`)
+      })
   }
 
   async use (req: Request, res: Response, next: NextFunction) {
-    let ltiMiddleware
-    if (this.ltiMiddleware === undefined) {
-      ltiMiddleware = await this.setLTIMiddleware()
-    } else {
-      ltiMiddleware = this.ltiMiddleware
-    }
-    ltiMiddleware(req, res, next)
+    if (this.ltiMiddleware !== undefined) this.ltiMiddleware(req, res, next)
   }
 }
