+++ ccm_web/server/src/lti/lti.module.ts
@@ -2,9 +2,11 @@ import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 
 import { LTIMiddleware } from './lti.middleware'
+import { LTIService } from './lti.service'
 
 @Module({
-  imports: [ConfigModule]
+  imports: [ConfigModule],
+  providers: [LTIService]
 })
 export class LtiModule implements NestModule {
   configure (consumer: MiddlewareConsumer) {
